+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Fixed
+
+- Redact authorization and cookie values from matcher failure diagnostics.
+
 ## 0.1.0 - 2026-07-26
 
 ### Added